Fix broken compilation on linux ( signed unsigned comparisson )
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@3345 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ libs/au/log/LogCentral.h
@@ -40,7 +40,7 @@ namespace au
                 
         SocketConnection * socket_connection;       // Socket connection with the logServer
         au::Cronometer time_since_last_connection;  // Cronometer with the time since last connection
-        int time_reconnect;                         // time for the next reconnection
+        size_t time_reconnect;                         // time for the next reconnection
         
         // Local file descriptor to write the log if not possible to connect
         FileDescriptor* local_file_descriptor;
